requirements.txt
discord.py
firebase-admin
python-dotenv
asyncpg
sqlalchemy[asyncio]
alembic
cryptography